solid designation component
